src/gcf_download_html_and_pdf/requirements.txt
Flask~=1.1.2
Pillow==9.3.0
pytesseract==0.3.10
pdf2image==1.16.3
numpy==1.22.0
pyarrow==11.0.0
# tesserocr==2.5.1
requests==2.28.0
gunicorn==20.1.0
markupsafe==2.0.1
beautifulsoup4==4.8.0
lxml==4.9.2
textract==1.6.5
python-dateutil==2.8.2